The Menus class was added to move the menus from the Main Class
